# api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API GPE",
    "version": "1.0.0",
    "description": "Documentação da API do sistema GPE desenvolvida em Node.js com Express e MySQL."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/registrar-presenca": {
      "post": {
        "tags": ["Professores"],
        "summary": "Registra presença de um aluno",
        "description": "Se o status for falta, o sistema também registra uma notificação e tenta enviar e-mail ao responsável.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "aluno_id": {
                    "type": "integer",
                    "example": 1
                  },
                  "professor_id": {
                    "type": "integer",
                    "example": 1
                  },
                  "data": {
                    "type": "string",
                    "example": "2026-04-10"
                  },
                  "status": {
                    "type": "string",
                    "enum": ["presente", "falta", "justificado"],
                    "example": "presente"
                  }
                },
                "required": ["aluno_id", "professor_id", "data", "status"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Presença registrada com sucesso"
          },
          "400": {
            "description": "Campos obrigatórios ausentes"
          },
          "500": {
            "description": "Erro ao registrar presença"
          }
        }
      }
    },
    "/api/presencas": {
      "get": {
        "tags": ["Professores","Gestores"],
        "summary": "Lista todas as presenças",
        "responses": {
          "200": {
            "description": "Lista retornada com sucesso"
          },
          "500": {
            "description": "Erro ao buscar presenças"
          }
        }
      }
    },
    "/api/presencas/{usuario_id}": {
      "get": {
        "tags": ["Professores"],
        "summary": "Busca as presenças de um aluno pelo usuario_id",
        "parameters": [
          {
            "name": "usuario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Presenças retornadas com sucesso"
          },
          "500": {
            "description": "Erro ao buscar presenças"
          }
        }
      }
    },
    "/api/presencas-filho/{usuario_id}": {
      "get": {
        "tags": ["Professores","Gestores"],
        "summary": "Busca as presenças do filho vinculado ao responsável",
        "parameters": [
          {
            "name": "usuario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Presenças do filho retornadas com sucesso"
          },
          "500": {
            "description": "Erro ao buscar presenças do filho"
          }
        }
      }
    },
    "/api/relatorio-frequencia": {
      "get": {
        "tags": ["Gestores"],
        "summary": "Gera relatório de frequência",
        "responses": {
          "200": {
            "description": "Relatório gerado com sucesso"
          },
          "500": {
            "description": "Erro ao gerar relatório"
          }
        }
      }
    }
  }
}